+++ HwProj.CoursesService/HwProj.CoursesService.API/Events/NewHomeworkTaskEvent.cs
@@ -6,7 +6,7 @@ namespace HwProj.CoursesService.API.Events
 {
     public class NewHomeworkTaskEvent : Event
     {
-        public NewHomeworkTaskEvent(string taskTitle, long taskId, DateTime? deadlineDate, string courseName, string[] studentIds, int courseId)
+        public NewHomeworkTaskEvent(string taskTitle, long taskId, DateTime? deadlineDate, string courseName, string[] studentIds, long courseId)
         {
             TaskTitle = taskTitle;
             TaskId = taskId;
@@ -19,7 +19,7 @@ namespace HwProj.CoursesService.API.Events
         public string TaskTitle { get; set; }
         public long TaskId { get; set; }
         public DateTime? DeadlineDate { get; set; }
-        public int CourseId { get; set; }
+        public long CourseId { get; set; }
         public string CourseName { get; set; }
         public string[] StudentIds { get; set; }
     }
